requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
colorama==0.4.6
et_xmlfile==2.0.0
lxml==5.3.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pywin32-ctypes==0.2.3
setuptools==75.6.0
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
xlwt==1.3.0
